完成链码同步

See merge request rtbasia/dif-client!31

+++ install-chaincode.js
@@ -6,25 +6,29 @@ let util = require('util')
 
 var log4js = require('log4js')
 var logger = log4js.getLogger('install-chaincode')
+let path = require('path')
 
-var installChaincode = async function () {
+var installChaincode = async function (ccName, ccVersion, isRemote) {
   // 准备安装的链码
-  fs.removeSync('./chaincode/build')
-  fs.ensureDirSync('./chaincode/build')
-  fs.copySync('./chaincode/node/.npmrc', './chaincode/build/.npmrc')
-  fs.copySync('./chaincode/node/chaincode.js', './chaincode/build/chaincode.js')
-  fs.copySync('./chaincode/node/dif.js', './chaincode/build/dif.js')
-  fs.copySync('./chaincode/node/package.json', './chaincode/build/package.json')
+  if (!isRemote) {
+    fs.removeSync('./chaincode/build')
+    fs.ensureDirSync('./chaincode/build')
+    fs.copySync('./chaincode/node/.npmrc', './chaincode/build/.npmrc')
+    fs.copySync('./chaincode/node/chaincode.js', './chaincode/build/chaincode.js')
+    fs.copySync('./chaincode/node/dif.js', './chaincode/build/dif.js')
+    fs.copySync('./chaincode/node/package.json', './chaincode/build/package.json')
+  }
 
   let client = await helper.getClient(true)
   let peers = helper.getOwnPeers(client)
 
+  let ccPath = path.join(__dirname, './chaincode/build')
   var request = {
     targets: peers,
-    chaincodePath: './chaincode/build',
-    chaincodeId: 'dif',
+    chaincodePath: ccPath,
+    chaincodeId: ccName,
     chaincodeType: 'node',
-    chaincodeVersion: 'v7' // TODO: 这些信息应该都要从服务器取得
+    chaincodeVersion: ccVersion // TODO: 这些信息应该都要从服务器取得
   }
 
   let results = await client.installChaincode(request, 180000)
@@ -37,7 +41,7 @@ var installChaincode = async function () {
     let oneGood = false
 
     if (proposalResponses && proposalResponses[i].response &&
-            proposalResponses[i].response.status === 200) {
+      proposalResponses[i].response.status === 200) {
       oneGood = true
       logger.info('install proposal was good')
     } else {
@@ -51,11 +55,13 @@ var installChaincode = async function () {
     logger.info(util.format(
       'Successfully sent install Proposal and received ProposalResponse: Status - %s',
       proposalResponses[0].response.status))
+    return {success: true, message: '安装成功'}
   } else {
     logger.error(
       'Failed to send install Proposal or receive valid response. Response null or status is not 200. exiting...'
     )
+    return {success: false, message: '安装失败'}
   }
 }
 
-installChaincode()
+exports.installChaincode = installChaincode
